pyproject.toml
[project]
name = "studyguru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.6.1",
    "duckduckgo-search>=8.1.1",
    "email-validator>=2.2.0",
    "fastapi>=0.131.0",
    "google-cloud-storage>=2.19.0",
    "httpx>=0.28.1",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.7",
    "lxml>=6.0.2",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic-settings>=2.11.0",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.3.0",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.36",
    "tenacity>=9.1.2",
    "uvicorn>=0.41.0",
]
